refactor cli run_single
